modified managers
